revert: retry opening file if there is an error, works around contention with antivirus software
The original commit arose from a misunderstanding of how opportunistic locking works, and the source code comment is inaccurate. In fact, antivirus software that uses opportunistic locking correctly does not cause contention.

The problem the original commit was intended to resolve is now known to be a defect in the caller, not in TDbf itself.

Using opportunistic locks to get out of the way if somebody wants the file:
https://blogs.msdn.microsoft.com/oldnewthing/20130415-00/?p=4663